refactoring($compile): OrderValidator 클래스로 변경
